Update debug explanation
I _expect_ that a debug set to 'false' for a specific form actually overrides the global settings, when set to 'true'. Couldn't find anything in the docs and the old explanation in config made it not completely clear. 

If this text change does not reflect what actually happens with the overrides, feel free to discard! If it does, this might be clearer.